Refactor trade config into ES module
