+++ server/src/trade/broker/providers/alpaca/alpaca.js
@@ -1,7 +1,7 @@
 'use strict';
 
 import axios from 'axios';
-import tradeConfig from '../../../trade.config.json';
+import tradeConfig from '../../../trade.config.js';
 import alpacaConfig from './alpaca.config.js';
 const BROKER = 'ALPACA';
