+++ server/src/trade/sell/sell.js
@@ -7,7 +7,7 @@ import chalk from 'chalk';
 import notifier from 'node-notifier';
 import asyncMap from 'async/map';
 import Broker from '../broker/Broker.js';
-import tradeConfig from '../trade.config.json';
+import tradeConfig from '../trade.config.js';
 import sellConfig from './sell.config.js';
 import { omitBlacklistedSecurities } from '../strategies/shared/common/common-evals.js';
 import { composeEvalFunctions, getSecurityData } from '../trade-utils.js';
