+++ server/src/trade/trade.config.js
@@ -1,0 +1,16 @@
+export default {
+    broker: 'ALPACA',
+    brokerProviders: {
+        ALPACA: 'alpaca'
+    },
+    whitelist: [],
+    blacklist: [
+        'PYPL'
+    ],
+    tradeAmount: 1.00,
+    tradeQty: 1,
+    maxSymbolsPerJob: 10,
+    tradeUnit: 'SHARES',
+    orderType: 'market',
+    timeInForce: 'gtc'
+};
