+++ test/unit/server/trade/trade.config.test.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import tradeConfig from '../../../../server/src/trade/trade.config.json';
+import tradeConfig from '../../../../server/src/trade/trade.config.js';
 import CONSTANTS from '../../../../server/src/constants.json';
 
 describe('Trade Config', () => {
